:art: refactor(NumberService): 클래스 구조 변경
- NumberService를 GameService와 LottoService로 분리
  - GameService : PurchaseAmount, RankStatistics
  - LottoService : Lottos, WinningNumbers
